05-upgraded-gvi-tool/requirements-v0.7.txt
numpy>=1.24
pandas>=2.0
opencv-python>=4.8
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
pillow>=10.0
transformers>=4.40
torch>=2.1
torchvision>=0.16
